+++ src/blitter/32bpp_base.cpp
@@ -18,12 +18,12 @@ void *Blitter_32bppBase::MoveTo(void *video, int x, int y)
 	return (uint32 *)video + x + y * _screen.pitch;
 }
 
-void Blitter_32bppBase::SetPixel(void *video, int x, int y, uint8 colour)
+void Blitter_32bppBase::SetPixel(void *video, int x, int y, uint16 colour)
 {
 	*((Colour *)video + x + y * _screen.pitch) = LookupColourInPalette(colour);
 }
 
-void Blitter_32bppBase::DrawLine(void *video, int x, int y, int x2, int y2, int screen_width, int screen_height, uint8 colour, int width, int dash)
+void Blitter_32bppBase::DrawLine(void *video, int x, int y, int x2, int y2, int screen_width, int screen_height, uint16 colour, int width, int dash)
 {
 	const Colour c = LookupColourInPalette(colour);
 	this->DrawLineGeneric(x, y, x2, y2, screen_width, screen_height, width, dash, [=](int x, int y) {
@@ -31,10 +31,10 @@ void Blitter_32bppBase::DrawLine(void *video, int x, int y, int x2, int y2, int 
 	});
 }
 
-void Blitter_32bppBase::DrawRect(void *video, int width, int height, uint8 colour)
+void Blitter_32bppBase::DrawRect(void *video, int width, int height, uint16 colour)
 {
 	Colour colour32 = LookupColourInPalette(colour);
-
+// 	printf("DrawRect %d, %d, %d\n", colour32.r, colour32.g, colour32.b);
 	do {
 		Colour *dst = (Colour *)video;
 		for (int i = width; i > 0; i--) {
