+++ src/gfx_type.h
@@ -223,7 +223,7 @@ struct SubSprite {
 	int left, top, right, bottom;
 };
 
-enum Colours : byte {
+enum Colours : uint8 {
 	COLOUR_BEGIN,
 	COLOUR_DARK_BLUE = COLOUR_BEGIN,
 	COLOUR_PALE_GREEN,
@@ -243,8 +243,9 @@ enum Colours : byte {
 	COLOUR_WHITE,
 	COLOUR_END,
 	INVALID_COLOUR = 0xFF,
+
 };
-template <> struct EnumPropsT<Colours> : MakeEnumPropsT<Colours, byte, COLOUR_BEGIN, COLOUR_END, INVALID_COLOUR, 8> {};
+template <> struct EnumPropsT<Colours> : MakeEnumPropsT<Colours, uint16, COLOUR_BEGIN, COLOUR_END, INVALID_COLOUR, 8> {};
 
 /** Colour of the strings, see _string_colourmap in table/string_colours.h or docs/ottd-colourtext-palette.png */
 enum TextColour {
