+++ src/livery.h
@@ -77,8 +77,8 @@ enum LiveryClass {
 /** Information about a particular livery. */
 struct Livery {
 	byte in_use;  ///< Bit 0 set if this livery should override the default livery first colour, Bit 1 for the second colour.
-	byte colour1; ///< First colour, for all vehicles.
-	byte colour2; ///< Second colour, for vehicles with 2CC support.
+	uint16 colour1; ///< First colour, for all vehicles.
+	uint16 colour2; ///< Second colour, for vehicles with 2CC support.
 };
 
 void ResetCompanyLivery(Company *c);
