+++ src/network/core/game_info.h
@@ -100,11 +100,11 @@ struct NetworkServerGameInfo {
 	std::string server_revision; ///< The version number the server is using (e.g.: 'r304' or 0.5.0)
 	bool dedicated;              ///< Is this a dedicated server?
 	bool use_password;           ///< Is this server passworded?
-	byte clients_on;             ///< Current count of clients on server
-	byte clients_max;            ///< Max clients allowed on server
-	byte companies_on;           ///< How many started companies do we have
-	byte companies_max;          ///< Max companies allowed on server
-	byte spectators_on;          ///< How many spectators do we have?
+	uint16 clients_on;             ///< Current count of clients on server
+	uint16 clients_max;            ///< Max clients allowed on server
+	uint16 companies_on;           ///< How many started companies do we have
+	uint16 companies_max;          ///< Max companies allowed on server
+	uint16 spectators_on;          ///< How many spectators do we have?
 	byte landscape;              ///< The used landscape
 	int gamescript_version;      ///< Version of the gamescript.
 	std::string gamescript_name; ///< Name of the gamescript.
